+++ tests/fixtures/projects/test_command_fail/Sources/main.xs
@@ -1,6 +1,6 @@
 // SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
 // SPDX-License-Identifier: MPL-2.0
 
-#[Test]
-fn broken( {
+fn main() -> Long {
+    return 0;
 }