+++ tests/fixtures/projects/test_command_invalid/Sources/Test/invalid_signature.xs
@@ -2,11 +2,5 @@
 // SPDX-License-Identifier: MPL-2.0
 
 #[Test]
-fn arithmetic_works() {
-    value: Long = 1 + 1;
-}
-
-#[Test]
-#[Ignore]
-fn future_test() {
+fn invalid_signature(value: Long) {
 }