+++ tests/fixtures/projects/test_command_should_panic/Sources/main.xs
@@ -1,12 +1,6 @@
 // SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
 // SPDX-License-Identifier: MPL-2.0
 
-#[Test]
-fn arithmetic_works() {
-    value: Long = 1 + 1;
-}
-
-#[Test]
-#[Ignore]
-fn future_test() {
+fn main() -> Long {
+    return 0;
 }