+++ tests/fixtures/projects/test_command_should_panic/xs.project.kts
@@ -1,12 +1,8 @@
 // SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
 // SPDX-License-Identifier: MPL-2.0
 
-#[Test]
-fn arithmetic_works() {
-    value: Long = 1 + 1;
-}
+project("ShouldPanicTestCommand", "BETA", "0.2.1")
 
-#[Test]
-#[Ignore]
-fn future_test() {
+source {
+  include("Sources")
 }